Make the increment and decrement logic more extensible.
The old logic was saying "if I'm moving from Juryo to Makuuchi do this, if I'm going the other way do this". Thing is if you add Makushita and Sandanme you get a combinatorial explosion with this pattern. What you really want is "if I'm leaving Makuuchi, do this. if I'm entering Juryo do this". Decouple, in other words.